+++ src/uint.rs
@@ -28,7 +28,6 @@
 //! ```
 //!
 
-use core::marker::PhantomData;
 use core::ops::{Add, BitAnd, BitOr, BitXor, Mul, Shl, Shr, Sub};
 use {Cmp, Equal, Greater, IsGreaterOrEqual, Len, Less, Logarithm2, NonZero, Ord, Pow, SquareRoot};
 
@@ -42,6 +41,8 @@ use private::{
 use private::{
     BitDiffOut, PrivateAndOut, PrivateCmpOut, PrivatePowOut, PrivateSubOut, PrivateXorOut, TrimOut,
 };
+
+use private::{Internal, InternalMarker};
 
 use consts::{U0, U1};
 use {Add1, Double, GrEq, Length, Log2, Or, Prod, Shleft, Shright, Sqrt, Square, Sub1, Sum};
@@ -149,16 +150,17 @@ impl Unsigned for UTerm {
 /// ```
 #[derive(Eq, PartialEq, Ord, PartialOrd, Clone, Copy, Hash, Debug, Default)]
 pub struct UInt<U, B> {
-    _marker: PhantomData<(U, B)>,
+    /// The more significant bits of `Self`.
+    pub(crate) msb: U,
+    /// The least significant bit of `Self`.
+    pub(crate) lsb: B,
 }
 
 impl<U: Unsigned, B: Bit> UInt<U, B> {
     /// Instantiates a singleton representing this unsigned integer.
     #[inline]
     pub fn new() -> UInt<U, B> {
-        UInt {
-            _marker: PhantomData,
-        }
+        UInt::default()
     }
 }
 
@@ -257,7 +259,7 @@ where
 {
     type Output = Add1<Length<U>>;
     fn len(&self) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+        self.msb.len() + B1
     }
 }
 
@@ -314,8 +316,8 @@ where
 /// `UTerm + U = U`
 impl<U: Unsigned> Add<U> for UTerm {
     type Output = U;
-    fn add(self, _: U) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn add(self, rhs: U) -> Self::Output {
+        rhs
     }
 }
 
@@ -333,8 +335,8 @@ where
     Ul: Add<Ur>,
 {
     type Output = UInt<Sum<Ul, Ur>, B0>;
-    fn add(self, _: UInt<Ur, B0>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn add(self, rhs: UInt<Ur, B0>) -> Self::Output {
+        UInt { msb: self.msb + rhs.msb, lsb: B0 }
     }
 }
 
@@ -344,8 +346,8 @@ where
     Ul: Add<Ur>,
 {
     type Output = UInt<Sum<Ul, Ur>, B1>;
-    fn add(self, _: UInt<Ur, B1>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn add(self, rhs: UInt<Ur, B1>) -> Self::Output {
+        UInt { msb: self.msb + rhs.msb, lsb: B1 }
     }
 }
 
@@ -355,8 +357,8 @@ where
     Ul: Add<Ur>,
 {
     type Output = UInt<Sum<Ul, Ur>, B1>;
-    fn add(self, _: UInt<Ur, B0>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn add(self, rhs: UInt<Ur, B0>) -> Self::Output {
+        UInt { msb: self.msb + rhs.msb, lsb: B1 }
     }
 }
 
@@ -367,8 +369,8 @@ where
     Sum<Ul, Ur>: Add<B1>,
 {
     type Output = UInt<Add1<Sum<Ul, Ur>>, B0>;
-    fn add(self, _: UInt<Ur, B1>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn add(self, rhs: UInt<Ur, B1>) -> Self::Output {
+        UInt { msb: self.msb + rhs.msb + B1, lsb: B0 }
     }
 }
 
@@ -437,14 +439,18 @@ where
     PrivateSubOut<UInt<Ul, Bl>, Ur>: Trim,
 {
     type Output = TrimOut<PrivateSubOut<UInt<Ul, Bl>, Ur>>;
-    fn sub(self, _: Ur) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn sub(self, rhs: Ur) -> Self::Output {
+        self.private_sub(rhs).trim()
     }
 }
 
 /// `U - UTerm = U`
 impl<U: Unsigned> PrivateSub<UTerm> for U {
     type Output = U;
+
+    fn private_sub(self, _: UTerm) -> Self::Output {
+        self
+    }
 }
 
 /// `UInt<Ul, B0> - UInt<Ur, B0> = UInt<Ul - Ur, B0>`
@@ -453,6 +459,10 @@ where
     Ul: PrivateSub<Ur>,
 {
     type Output = UInt<PrivateSubOut<Ul, Ur>, B0>;
+
+    fn private_sub(self, rhs: UInt<Ur, B0>) -> Self::Output {
+        UInt { msb: self.msb.private_sub(rhs.msb), lsb: B0 }
+    }
 }
 
 /// `UInt<Ul, B0> - UInt<Ur, B1> = UInt<(Ul - Ur) - B1, B1>`
@@ -462,6 +472,10 @@ where
     PrivateSubOut<Ul, Ur>: Sub<B1>,
 {
     type Output = UInt<Sub1<PrivateSubOut<Ul, Ur>>, B1>;
+
+    fn private_sub(self, rhs: UInt<Ur, B1>) -> Self::Output {
+        UInt { msb: self.msb.private_sub(rhs.msb) - B1, lsb: B1 }
+    }
 }
 
 /// `UInt<Ul, B1> - UInt<Ur, B0> = UInt<Ul - Ur, B1>`
@@ -470,6 +484,10 @@ where
     Ul: PrivateSub<Ur>,
 {
     type Output = UInt<PrivateSubOut<Ul, Ur>, B1>;
+
+    fn private_sub(self, rhs: UInt<Ur, B0>) -> Self::Output {
+        UInt { msb: self.msb.private_sub(rhs.msb), lsb: B1 }
+    }
 }
 
 /// `UInt<Ul, B1> - UInt<Ur, B1> = UInt<Ul - Ur, B0>`
@@ -478,6 +496,10 @@ where
     Ul: PrivateSub<Ur>,
 {
     type Output = UInt<PrivateSubOut<Ul, Ur>, B0>;
+
+    fn private_sub(self, rhs: UInt<Ur, B1>) -> Self::Output {
+        UInt { msb: self.msb.private_sub(rhs.msb), lsb: B0 }
+    }
 }
 
 // ---------------------------------------------------------------------------------------
@@ -499,19 +521,27 @@ where
     PrivateAndOut<UInt<Ul, Bl>, Ur>: Trim,
 {
     type Output = TrimOut<PrivateAndOut<UInt<Ul, Bl>, Ur>>;
-    fn bitand(self, _: Ur) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn bitand(self, rhs: Ur) -> Self::Output {
+        self.private_and(rhs).trim()
     }
 }
 
 /// `UTerm & X = UTerm`
 impl<U: Unsigned> PrivateAnd<U> for UTerm {
     type Output = UTerm;
+
+    fn private_and(self, _: U) -> Self::Output {
+        UTerm
+    }
 }
 
 /// `X & UTerm = UTerm`
 impl<B: Bit, U: Unsigned> PrivateAnd<UTerm> for UInt<U, B> {
     type Output = UTerm;
+
+    fn private_and(self, _: UTerm) -> Self::Output {
+        UTerm
+    }
 }
 
 /// `UInt<Ul, B0> & UInt<Ur, B0> = UInt<Ul & Ur, B0>`
@@ -520,6 +550,10 @@ where
     Ul: PrivateAnd<Ur>,
 {
     type Output = UInt<PrivateAndOut<Ul, Ur>, B0>;
+
+    fn private_and(self, rhs: UInt<Ur, B0>) -> Self::Output {
+        UInt { msb: self.msb.private_and(rhs.msb), lsb: B0 }
+    }
 }
 
 /// `UInt<Ul, B0> & UInt<Ur, B1> = UInt<Ul & Ur, B0>`
@@ -528,6 +562,10 @@ where
     Ul: PrivateAnd<Ur>,
 {
     type Output = UInt<PrivateAndOut<Ul, Ur>, B0>;
+
+    fn private_and(self, rhs: UInt<Ur, B1>) -> Self::Output {
+        UInt { msb: self.msb.private_and(rhs.msb), lsb: B0 }
+    }
 }
 
 /// `UInt<Ul, B1> & UInt<Ur, B0> = UInt<Ul & Ur, B0>`
@@ -536,6 +574,10 @@ where
     Ul: PrivateAnd<Ur>,
 {
     type Output = UInt<PrivateAndOut<Ul, Ur>, B0>;
+
+    fn private_and(self, rhs: UInt<Ur, B0>) -> Self::Output {
+        UInt { msb: self.msb.private_and(rhs.msb), lsb: B0 }
+    }
 }
 
 /// `UInt<Ul, B1> & UInt<Ur, B1> = UInt<Ul & Ur, B1>`
@@ -544,6 +586,10 @@ where
     Ul: PrivateAnd<Ur>,
 {
     type Output = UInt<PrivateAndOut<Ul, Ur>, B1>;
+
+    fn private_and(self, rhs: UInt<Ur, B1>) -> Self::Output {
+        UInt { msb: self.msb.private_and(rhs.msb), lsb: B1 }
+    }
 }
 
 // ---------------------------------------------------------------------------------------
@@ -552,8 +598,8 @@ where
 /// `UTerm | X = X`
 impl<U: Unsigned> BitOr<U> for UTerm {
     type Output = U;
-    fn bitor(self, _: U) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn bitor(self, rhs: U) -> Self::Output {
+        rhs
     }
 }
 
@@ -571,8 +617,8 @@ where
     Ul: BitOr<Ur>,
 {
     type Output = UInt<<Ul as BitOr<Ur>>::Output, B0>;
-    fn bitor(self, _: UInt<Ur, B0>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn bitor(self, rhs: UInt<Ur, B0>) -> Self::Output {
+        UInt { msb: self.msb.bitor(rhs.msb), lsb: B0 }
     }
 }
 
@@ -582,8 +628,8 @@ where
     Ul: BitOr<Ur>,
 {
     type Output = UInt<Or<Ul, Ur>, B1>;
-    fn bitor(self, _: UInt<Ur, B1>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn bitor(self, rhs: UInt<Ur, B1>) -> Self::Output {
+        UInt { msb: self.msb.bitor(rhs.msb), lsb: self.lsb.bitor(rhs.lsb) }
     }
 }
 
@@ -593,8 +639,8 @@ where
     Ul: BitOr<Ur>,
 {
     type Output = UInt<Or<Ul, Ur>, B1>;
-    fn bitor(self, _: UInt<Ur, B0>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn bitor(self, rhs: UInt<Ur, B0>) -> Self::Output {
+        UInt { msb: self.msb.bitor(rhs.msb), lsb: self.lsb.bitor(rhs.lsb) }
     }
 }
 
@@ -604,8 +650,8 @@ where
     Ul: BitOr<Ur>,
 {
     type Output = UInt<Or<Ul, Ur>, B1>;
-    fn bitor(self, _: UInt<Ur, B1>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn bitor(self, rhs: UInt<Ur, B1>) -> Self::Output {
+        UInt { msb: self.msb.bitor(rhs.msb), lsb: self.lsb.bitor(rhs.lsb) }
     }
 }
 
@@ -615,8 +661,8 @@ where
 /// 0 ^ X = X
 impl<Ur: Unsigned> BitXor<Ur> for UTerm {
     type Output = Ur;
-    fn bitxor(self, _: Ur) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn bitxor(self, rhs: Ur) -> Self::Output {
+        rhs
     }
 }
 
@@ -628,19 +674,27 @@ where
     PrivateXorOut<UInt<Ul, Bl>, Ur>: Trim,
 {
     type Output = TrimOut<PrivateXorOut<UInt<Ul, Bl>, Ur>>;
-    fn bitxor(self, _: Ur) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn bitxor(self, rhs: Ur) -> Self::Output {
+        self.private_xor(rhs).trim()
     }
 }
 
 /// `UTerm ^ X = X`
 impl<U: Unsigned> PrivateXor<U> for UTerm {
     type Output = U;
+
+    fn private_xor(self, rhs: U) -> Self::Output {
+        rhs
+    }
 }
 
 /// `X ^ UTerm = X`
 impl<B: Bit, U: Unsigned> PrivateXor<UTerm> for UInt<U, B> {
     type Output = Self;
+
+    fn private_xor(self, _: UTerm) -> Self::Output {
+        self
+    }
 }
 
 /// `UInt<Ul, B0> ^ UInt<Ur, B0> = UInt<Ul ^ Ur, B0>`
@@ -649,6 +703,10 @@ where
     Ul: PrivateXor<Ur>,
 {
     type Output = UInt<PrivateXorOut<Ul, Ur>, B0>;
+
+    fn private_xor(self, rhs: UInt<Ur, B0>) -> Self::Output {
+        UInt { msb: self.msb.private_xor(rhs.msb), lsb: B0 }
+    }
 }
 
 /// `UInt<Ul, B0> ^ UInt<Ur, B1> = UInt<Ul ^ Ur, B1>`
@@ -657,6 +715,10 @@ where
     Ul: PrivateXor<Ur>,
 {
     type Output = UInt<PrivateXorOut<Ul, Ur>, B1>;
+
+    fn private_xor(self, rhs: UInt<Ur, B1>) -> Self::Output {
+        UInt { msb: self.msb.private_xor(rhs.msb), lsb: B1 }
+    }
 }
 
 /// `UInt<Ul, B1> ^ UInt<Ur, B0> = UInt<Ul ^ Ur, B1>`
@@ -665,6 +727,10 @@ where
     Ul: PrivateXor<Ur>,
 {
     type Output = UInt<PrivateXorOut<Ul, Ur>, B1>;
+
+    fn private_xor(self, rhs: UInt<Ur, B0>) -> Self::Output {
+        UInt { msb: self.msb.private_xor(rhs.msb), lsb: B1 }
+    }
 }
 
 /// `UInt<Ul, B1> ^ UInt<Ur, B1> = UInt<Ul ^ Ur, B0>`
@@ -673,6 +739,10 @@ where
     Ul: PrivateXor<Ur>,
 {
     type Output = UInt<PrivateXorOut<Ul, Ur>, B0>;
+
+    fn private_xor(self, rhs: UInt<Ur, B1>) -> Self::Output {
+        UInt { msb: self.msb.private_xor(rhs.msb), lsb: B0 }
+    }
 }
 
 // ---------------------------------------------------------------------------------------
@@ -733,8 +803,8 @@ where
     UInt<UInt<U, B>, B0>: Shl<Sub1<UInt<Ur, Br>>>,
 {
     type Output = Shleft<UInt<UInt<U, B>, B0>, Sub1<UInt<Ur, Br>>>;
-    fn shl(self, _: UInt<Ur, Br>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn shl(self, rhs: UInt<Ur, Br>) -> Self::Output {
+        (UInt { msb: self, lsb: B0 }).shl(rhs - B1)
     }
 }
 
@@ -785,7 +855,7 @@ impl<U: Unsigned, B: Bit> Shr<B0> for UInt<U, B> {
 impl<U: Unsigned, B: Bit> Shr<B1> for UInt<U, B> {
     type Output = U;
     fn shr(self, _: B1) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+        self.msb
     }
 }
 
@@ -796,8 +866,8 @@ where
     U: Shr<Sub1<UInt<Ur, Br>>>,
 {
     type Output = Shright<U, Sub1<UInt<Ur, Br>>>;
-    fn shr(self, _: UInt<Ur, Br>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn shr(self, rhs: UInt<Ur, Br>) -> Self::Output {
+        self.msb.shr(rhs - B1)
     }
 }
 
@@ -858,8 +928,8 @@ where
     Ul: Mul<UInt<Ur, B>>,
 {
     type Output = UInt<Prod<Ul, UInt<Ur, B>>, B0>;
-    fn mul(self, _: UInt<Ur, B>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn mul(self, rhs: UInt<Ur, B>) -> Self::Output {
+        UInt { msb: self.msb * rhs, lsb: B0 }
     }
 }
 
@@ -870,8 +940,8 @@ where
     UInt<Prod<Ul, UInt<Ur, B>>, B0>: Add<UInt<Ur, B>>,
 {
     type Output = Sum<UInt<Prod<Ul, UInt<Ur, B>>, B0>, UInt<Ur, B>>;
-    fn mul(self, _: UInt<Ur, B>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn mul(self, rhs: UInt<Ur, B>) -> Self::Output {
+        UInt { msb: self.msb * rhs, lsb: B0 } + rhs
     }
 }
 
@@ -881,16 +951,28 @@ where
 /// Zero == Zero
 impl Cmp<UTerm> for UTerm {
     type Output = Equal;
+
+    fn compare<IM: InternalMarker>(&self, _: &UTerm) -> Self::Output {
+        Equal
+    }
 }
 
 /// Nonzero > Zero
 impl<U: Unsigned, B: Bit> Cmp<UTerm> for UInt<U, B> {
     type Output = Greater;
+
+    fn compare<IM: InternalMarker>(&self, _: &UTerm) -> Self::Output {
+        Greater
+    }
 }
 
 /// Zero < Nonzero
 impl<U: Unsigned, B: Bit> Cmp<UInt<U, B>> for UTerm {
     type Output = Less;
+
+    fn compare<IM: InternalMarker>(&self, _: &UInt<U, B>) -> Self::Output {
+        Less
+    }
 }
 
 /// `UInt<Ul, B0>` cmp with `UInt<Ur, B0>`: `SoFar` is `Equal`
@@ -899,6 +981,10 @@ where
     Ul: PrivateCmp<Ur, Equal>,
 {
     type Output = PrivateCmpOut<Ul, Ur, Equal>;
+
+    fn compare<IM: InternalMarker>(&self, rhs: &UInt<Ur, B0>) -> Self::Output {
+        self.msb.private_cmp(&rhs.msb, Equal)
+    }
 }
 
 /// `UInt<Ul, B1>` cmp with `UInt<Ur, B1>`: `SoFar` is `Equal`
@@ -907,6 +993,10 @@ where
     Ul: PrivateCmp<Ur, Equal>,
 {
     type Output = PrivateCmpOut<Ul, Ur, Equal>;
+
+    fn compare<IM: InternalMarker>(&self, rhs: &UInt<Ur, B1>) -> Self::Output {
+        self.msb.private_cmp(&rhs.msb, Equal)
+    }
 }
 
 /// `UInt<Ul, B0>` cmp with `UInt<Ur, B1>`: `SoFar` is `Less`
@@ -915,6 +1005,10 @@ where
     Ul: PrivateCmp<Ur, Less>,
 {
     type Output = PrivateCmpOut<Ul, Ur, Less>;
+
+    fn compare<IM: InternalMarker>(&self, rhs: &UInt<Ur, B1>) -> Self::Output {
+        self.msb.private_cmp(&rhs.msb, Less)
+    }
 }
 
 /// `UInt<Ul, B1>` cmp with `UInt<Ur, B0>`: `SoFar` is `Greater`
@@ -923,6 +1017,10 @@ where
     Ul: PrivateCmp<Ur, Greater>,
 {
     type Output = PrivateCmpOut<Ul, Ur, Greater>;
+
+    fn compare<IM: InternalMarker>(&self, rhs: &UInt<Ur, B0>) -> Self::Output {
+        self.msb.private_cmp(&rhs.msb, Greater)
+    }
 }
 
 /// Comparing non-terimal bits, with both having bit `B0`.
@@ -935,6 +1033,10 @@ where
     Ul: PrivateCmp<Ur, SoFar>,
 {
     type Output = PrivateCmpOut<Ul, Ur, SoFar>;
+
+    fn private_cmp(&self, rhs: &UInt<Ur, B0>, so_far: SoFar) -> Self::Output {
+        self.msb.private_cmp(&rhs.msb, so_far)
+    }
 }
 
 /// Comparing non-terimal bits, with both having bit `B1`.
@@ -947,6 +1049,10 @@ where
     Ul: PrivateCmp<Ur, SoFar>,
 {
     type Output = PrivateCmpOut<Ul, Ur, SoFar>;
+
+    fn private_cmp(&self, rhs: &UInt<Ur, B1>, so_far: SoFar) -> Self::Output {
+        self.msb.private_cmp(&rhs.msb, so_far)
+    }
 }
 
 /// Comparing non-terimal bits, with `Lhs` having bit `B0` and `Rhs` having bit `B1`.
@@ -959,6 +1065,10 @@ where
     Ul: PrivateCmp<Ur, Less>,
 {
     type Output = PrivateCmpOut<Ul, Ur, Less>;
+
+    fn private_cmp(&self, rhs: &UInt<Ur, B1>, _: SoFar) -> Self::Output {
+        self.msb.private_cmp(&rhs.msb, Less)
+    }
 }
 
 /// Comparing non-terimal bits, with `Lhs` having bit `B1` and `Rhs` having bit `B0`.
@@ -971,21 +1081,37 @@ where
     Ul: PrivateCmp<Ur, Greater>,
 {
     type Output = PrivateCmpOut<Ul, Ur, Greater>;
+
+    fn private_cmp(&self, rhs: &UInt<Ur, B0>, _: SoFar) -> Self::Output {
+        self.msb.private_cmp(&rhs.msb, Greater)
+    }
 }
 
 /// Got to the end of just the `Lhs`. It's `Less`.
 impl<U: Unsigned, B: Bit, SoFar: Ord> PrivateCmp<UInt<U, B>, SoFar> for UTerm {
     type Output = Less;
+
+    fn private_cmp(&self, _: &UInt<U, B>, _: SoFar) -> Self::Output {
+        Less
+    }
 }
 
 /// Got to the end of just the `Rhs`. `Lhs` is `Greater`.
 impl<U: Unsigned, B: Bit, SoFar: Ord> PrivateCmp<UTerm, SoFar> for UInt<U, B> {
     type Output = Greater;
+
+    fn private_cmp(&self, _: &UTerm, _: SoFar) -> Self::Output {
+        Greater
+    }
 }
 
 /// Got to the end of both! Return `SoFar`
 impl<SoFar: Ord> PrivateCmp<UTerm, SoFar> for UTerm {
     type Output = SoFar;
+
+    fn private_cmp(&self, _: &UTerm, so_far: SoFar) -> Self::Output {
+        so_far
+    }
 }
 
 // ---------------------------------------------------------------------------------------
@@ -1029,13 +1155,17 @@ where
     X: PrivatePow<U1, N>,
 {
     type Output = PrivatePowOut<X, U1, N>;
-    fn powi(self, _: N) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn powi(self, n: N) -> Self::Output {
+        self.private_pow(U1::new(), n)
     }
 }
 
 impl<Y: Unsigned, X: Unsigned> PrivatePow<Y, U0> for X {
     type Output = Y;
+
+    fn private_pow(self, y: Y, _: U0) -> Self::Output {
+        y
+    }
 }
 
 impl<Y: Unsigned, X: Unsigned> PrivatePow<Y, U1> for X
@@ -1043,6 +1173,10 @@ where
     X: Mul<Y>,
 {
     type Output = Prod<X, Y>;
+
+    fn private_pow(self, y: Y, _: U1) -> Self::Output {
+        self * y
+    }
 }
 
 /// N is even
@@ -1052,6 +1186,10 @@ where
     Square<X>: PrivatePow<Y, UInt<U, B>>,
 {
     type Output = PrivatePowOut<Square<X>, Y, UInt<U, B>>;
+
+    fn private_pow(self, y: Y, n: UInt<UInt<U, B>, B0>) -> Self::Output {
+        (self * self).private_pow(y, n.msb)
+    }
 }
 
 /// N is odd
@@ -1061,6 +1199,10 @@ where
     Square<X>: PrivatePow<Prod<X, Y>, UInt<U, B>>,
 {
     type Output = PrivatePowOut<Square<X>, Prod<X, Y>, UInt<U, B>>;
+
+    fn private_pow(self, y: Y, n: UInt<UInt<U, B>, B1>) -> Self::Output {
+        (self * self).private_pow(self * y, n.msb)
+    }
 }
 
 // -----------------------------------------
@@ -1070,28 +1212,46 @@ where
 pub trait GetBit<I> {
     #[allow(missing_docs)]
     type Output;
+
+    #[doc(hidden)]
+    fn get_bit<IM: InternalMarker>(&self, &I) -> Self::Output;
 }
 
 #[allow(missing_docs)]
 pub type GetBitOut<N, I> = <N as GetBit<I>>::Output;
 
 // Base case
-impl<Un, Bn> GetBit<U0> for UInt<Un, Bn> {
+impl<Un, Bn> GetBit<U0> for UInt<Un, Bn>
+where
+    Bn: Copy
+{
     type Output = Bn;
+
+    fn get_bit<IM: InternalMarker>(&self, _: &U0) -> Self::Output {
+        self.lsb
+    }
 }
 
 // Recursion case
 impl<Un, Bn, Ui, Bi> GetBit<UInt<Ui, Bi>> for UInt<Un, Bn>
 where
-    UInt<Ui, Bi>: Sub<B1>,
+    UInt<Ui, Bi>: Copy + Sub<B1>,
     Un: GetBit<Sub1<UInt<Ui, Bi>>>,
 {
     type Output = GetBitOut<Un, Sub1<UInt<Ui, Bi>>>;
+
+    fn get_bit<IM: InternalMarker>(&self, i: &UInt<Ui, Bi>) -> Self::Output {
+        self.msb.get_bit::<Internal>(&(*i - B1))
+    }
 }
 
 // Ran out of bits
 impl<I> GetBit<I> for UTerm {
     type Output = B0;
+
+    fn get_bit<IM: InternalMarker>(&self, _: &I) -> Self::Output {
+        B0
+    }
 }
 
 #[test]
@@ -1115,6 +1275,9 @@ fn test_get_bit() {
 pub trait SetBit<I, B> {
     #[allow(missing_docs)]
     type Output;
+
+    #[doc(hidden)]
+    fn set_bit<IM: InternalMarker>(self, I, B) -> Self::Output;
 }
 /// Alias for the result of calling `SetBit`: `SetBitOut<N, I, B> = <N as SetBit<I, B>>::Output`.
 pub type SetBitOut<N, I, B> = <N as SetBit<I, B>>::Output;
@@ -1128,11 +1291,19 @@ where
     PrivateSetBitOut<N, I, B>: Trim,
 {
     type Output = TrimOut<PrivateSetBitOut<N, I, B>>;
+
+    fn set_bit<IM: InternalMarker>(self, i: I, b: B) -> Self::Output {
+        self.private_set_bit(i, b).trim()
+    }
 }
 
 // Base case
 impl<Un, Bn, B> PrivateSetBit<U0, B> for UInt<Un, Bn> {
     type Output = UInt<Un, B>;
+
+    fn private_set_bit(self, _: U0, b: B) -> Self::Output {
+        UInt { msb: self.msb, lsb: b }
+    }
 }
 
 // Recursion case
@@ -1142,11 +1313,19 @@ where
     Un: PrivateSetBit<Sub1<UInt<Ui, Bi>>, B>,
 {
     type Output = UInt<PrivateSetBitOut<Un, Sub1<UInt<Ui, Bi>>, B>, Bn>;
+
+    fn private_set_bit(self, i: UInt<Ui, Bi>, b: B) -> Self::Output {
+        UInt { msb: self.msb.private_set_bit(i - B1, b), lsb: self.lsb }
+    }
 }
 
 // Ran out of bits, setting B0
 impl<I> PrivateSetBit<I, B0> for UTerm {
     type Output = UTerm;
+
+    fn private_set_bit(self, _: I, _: B0) -> Self::Output {
+        UTerm
+    }
 }
 
 // Ran out of bits, setting B1
@@ -1155,6 +1334,10 @@ where
     U1: Shl<I>,
 {
     type Output = Shleft<U1, I>;
+
+    fn private_set_bit(self, i: I, _: B1) -> Self::Output {
+        <U1 as Shl<I>>::shl(U1::new(), i)
+    }
 }
 
 #[test]
@@ -1252,8 +1435,14 @@ where
     (): PrivateDiv<UInt<Ul, Bl>, UInt<Ur, Br>, U0, U0, Sub1<Length<UInt<Ul, Bl>>>>,
 {
     type Output = PrivateDivQuot<UInt<Ul, Bl>, UInt<Ur, Br>, U0, U0, Sub1<Length<UInt<Ul, Bl>>>>;
-    fn div(self, _: UInt<Ur, Br>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn div(self, rhs: UInt<Ur, Br>) -> Self::Output {
+        ().private_div_quotient(
+            self,
+            rhs,
+            U0::new(),
+            U0::new(),
+            self.len() - B1,
+        )
     }
 }
 
@@ -1277,8 +1466,8 @@ where
     (): PrivateDiv<UInt<Ul, Bl>, UInt<Ur, Br>, U0, U0, Sub1<Length<UInt<Ul, Bl>>>>,
 {
     type Output = PrivateDivRem<UInt<Ul, Bl>, UInt<Ur, Br>, U0, U0, Sub1<Length<UInt<Ul, Bl>>>>;
-    fn rem(self, _: UInt<Ur, Br>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn rem(self, rhs: UInt<Ur, Br>) -> Self::Output {
+        ().private_div_remainder(self, rhs, UTerm, UTerm, self.len() - B1)
     }
 }
 
@@ -1318,6 +1507,35 @@ where
         I,
         Compare<TrimOut<UInt<UTerm, GetBitOut<N, I>>>, D>,
     >;
+
+    fn private_div_quotient(self, n: N, d: D, q: Q, _: U0, i: I) -> Self::Quotient
+    where
+    {
+        let r = (UInt { msb: UTerm, lsb: n.get_bit::<Internal>(&i) }).trim();
+        let r_cmp_d = r.compare::<Internal>(&d);
+        ().private_div_if_quotient(
+            n,
+            d,
+            q,
+            r,
+            i,
+            r_cmp_d
+        )
+    }
+
+    fn private_div_remainder(self, n: N, d: D, q: Q, _: U0, i: I) -> Self::Remainder
+    {
+        let r = (UInt { msb: UTerm, lsb: n.get_bit::<Internal>(&i) }).trim();
+        let r_cmp_d = r.compare::<Internal>(&d);
+        ().private_div_if_remainder(
+            n,
+            d,
+            q,
+            r,
+            i,
+            r_cmp_d
+        )
+    }
 }
 
 // R > 0: We perform R <<= 1 and R[0] = N[i], then call out to PrivateDivIf for the if statement
@@ -1350,6 +1568,32 @@ where
         I,
         Compare<UInt<UInt<Ur, Br>, GetBitOut<N, I>>, D>,
     >;
+
+    fn private_div_quotient(self, n: N, d: D, q: Q, r: UInt<Ur, Br>, i: I) -> Self::Quotient {
+        let r = UInt { msb: r, lsb: n.get_bit::<Internal>(&i) };
+        let r_cmp_d = r.compare::<Internal>(&d);
+        ().private_div_if_quotient(
+            n,
+            d,
+            q,
+            r,
+            i,
+            r_cmp_d
+        )
+    }
+
+    fn private_div_remainder(self, n: N, d: D, q: Q, r: UInt<Ur, Br>, i: I) -> Self::Remainder {
+        let r = UInt { msb: r, lsb: n.get_bit::<Internal>(&i) };
+        let r_cmp_d = r.compare::<Internal>(&d);
+        ().private_div_if_remainder(
+            n,
+            d,
+            q,
+            r,
+            i,
+            r_cmp_d
+        )
+    }
 }
 
 // -----------------------------------------
@@ -1365,24 +1609,49 @@ where
 {
     type Quotient = PrivateDivQuot<N, D, Q, R, Sub1<UInt<Ui, Bi>>>;
     type Remainder = PrivateDivRem<N, D, Q, R, Sub1<UInt<Ui, Bi>>>;
+
+    fn private_div_if_quotient(self, n: N, d: D, q: Q, r: R, i: UInt<Ui, Bi>, _: Less) -> Self::Quotient
+    where
+    {
+        ().private_div_quotient(n, d, q, r, i - B1)
+    }
+
+    fn private_div_if_remainder(self, n: N, d: D, q: Q, r: R, i: UInt<Ui, Bi>, _: Less) -> Self::Remainder
+    where
+    {
+        ().private_div_remainder(n, d, q, r, i - B1)
+    }
 }
 
 // R == D, I > 0, we set R = 0, Q[I] = 1 and recurse
 impl<N, D, Q, R, Ui, Bi> PrivateDivIf<N, D, Q, R, UInt<Ui, Bi>, Equal> for ()
 where
-    UInt<Ui, Bi>: Sub<B1>,
+    UInt<Ui, Bi>: Copy + Sub<B1>,
     Q: SetBit<UInt<Ui, Bi>, B1>,
     (): PrivateDiv<N, D, SetBitOut<Q, UInt<Ui, Bi>, B1>, U0, Sub1<UInt<Ui, Bi>>>,
 {
     type Quotient = PrivateDivQuot<N, D, SetBitOut<Q, UInt<Ui, Bi>, B1>, U0, Sub1<UInt<Ui, Bi>>>;
     type Remainder = PrivateDivRem<N, D, SetBitOut<Q, UInt<Ui, Bi>, B1>, U0, Sub1<UInt<Ui, Bi>>>;
+
+    fn private_div_if_quotient(self, n: N, d: D, q: Q, _: R, i: UInt<Ui, Bi>, _: Equal) -> Self::Quotient
+    where
+    {
+        ().private_div_quotient(n, d, q.set_bit::<Internal>(i, B1), U0::new(), i - B1)
+    }
+
+    fn private_div_if_remainder(self, n: N, d: D, q: Q, _: R, i: UInt<Ui, Bi>, _: Equal) -> Self::Remainder
+    where
+    {
+        ().private_div_remainder(n, d, q.set_bit::<Internal>(i, B1), U0::new(), i - B1)
+    }
 }
 
 use Diff;
 // R > D, I > 0, we set R -= D, Q[I] = 1 and recurse
 impl<N, D, Q, R, Ui, Bi> PrivateDivIf<N, D, Q, R, UInt<Ui, Bi>, Greater> for ()
 where
-    UInt<Ui, Bi>: Sub<B1>,
+    D: Copy,
+    UInt<Ui, Bi>: Copy + Sub<B1>,
     R: Sub<D>,
     Q: SetBit<UInt<Ui, Bi>, B1>,
     (): PrivateDiv<N, D, SetBitOut<Q, UInt<Ui, Bi>, B1>, Diff<R, D>, Sub1<UInt<Ui, Bi>>>,
@@ -1391,12 +1660,32 @@ where
         PrivateDivQuot<N, D, SetBitOut<Q, UInt<Ui, Bi>, B1>, Diff<R, D>, Sub1<UInt<Ui, Bi>>>;
     type Remainder =
         PrivateDivRem<N, D, SetBitOut<Q, UInt<Ui, Bi>, B1>, Diff<R, D>, Sub1<UInt<Ui, Bi>>>;
+
+    fn private_div_if_quotient(self, n: N, d: D, q: Q, r: R, i: UInt<Ui, Bi>, _: Greater) -> Self::Quotient
+    where
+    {
+        ().private_div_quotient(n, d, q.set_bit::<Internal>(i, B1), r - d, i - B1)
+    }
+
+    fn private_div_if_remainder(self, n: N, d: D, q: Q, r: R, i: UInt<Ui, Bi>, _: Greater) -> Self::Remainder
+    where
+    {
+        ().private_div_remainder(n, d, q.set_bit::<Internal>(i, B1), r - d, i - B1)
+    }
 }
 
 // R < D, I == 0: we do nothing, and return
 impl<N, D, Q, R> PrivateDivIf<N, D, Q, R, U0, Less> for () {
     type Quotient = Q;
     type Remainder = R;
+
+    fn private_div_if_quotient(self, _: N, _: D, q: Q, _: R, _: U0, _: Less) -> Self::Quotient {
+        q
+    }
+
+    fn private_div_if_remainder(self, _: N, _: D, _: Q, r: R, _: U0, _: Less) -> Self::Remainder {
+        r
+    }
 }
 
 // R == D, I == 0: we set R = 0, Q[I] = 1, and return
@@ -1406,6 +1695,14 @@ where
 {
     type Quotient = SetBitOut<Q, U0, B1>;
     type Remainder = U0;
+
+    fn private_div_if_quotient(self, _: N, _: D, q: Q, _: R, i: U0, _: Equal) -> Self::Quotient {
+        q.set_bit::<Internal>(i, B1)
+    }
+
+    fn private_div_if_remainder(self, _: N, _: D, _: Q, _: R, i: U0, _: Equal) -> Self::Remainder {
+        i
+    }
 }
 
 // R > D, I == 0: We set R -= D, Q[I] = 1, and return
@@ -1416,6 +1713,14 @@ where
 {
     type Quotient = SetBitOut<Q, U0, B1>;
     type Remainder = Diff<R, D>;
+
+    fn private_div_if_quotient(self, _: N, _: D, q: Q, _: R, i: U0, _: Greater) -> Self::Quotient {
+        q.set_bit::<Internal>(i, B1)
+    }
+
+    fn private_div_if_remainder(self, _: N, d: D, _: Q, r: R, _: U0, _: Greater) -> Self::Remainder {
+        r - d
+    }
 }
 
 // -----------------------------------------
@@ -1434,8 +1739,8 @@ where
     UInt<Ul, Bl>: Div<UInt<Ur, Br>> + Rem<UInt<Ur, Br>, Output = U0>,
 {
     type Output = Quot<UInt<Ul, Bl>, UInt<Ur, Br>>;
-    fn partial_div(self, _: UInt<Ur, Br>) -> Self::Output {
-        unsafe { ::core::mem::uninitialized() }
+    fn partial_div(self, rhs: UInt<Ur, Br>) -> Self::Output {
+        self / rhs
     }
 }